+++ packages/gatsby-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.8](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.7...@honeybadger-io/gatsby-plugin-honeybadger@6.0.8) (2023-08-21)
+
+**Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger
+
+
+
+
+
+## [6.0.7](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.6...@honeybadger-io/gatsby-plugin-honeybadger@6.0.7) (2023-08-07)
+
+**Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger
+
+
+
+
+
 ## [6.0.6](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.5...@honeybadger-io/gatsby-plugin-honeybadger@6.0.6) (2023-07-25)
 
 **Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger