+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.7.2](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/nextjs@5.7.1...@honeybadger-io/nextjs@5.7.2) (2023-08-21)
+
+**Note:** Version bump only for package @honeybadger-io/nextjs
+
+
+
+
+
+## [5.7.1](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/nextjs@5.7.0...@honeybadger-io/nextjs@5.7.1) (2023-08-07)
+
+**Note:** Version bump only for package @honeybadger-io/nextjs
+
+
+
+
+
 ## [5.7.0](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/nextjs@5.6.3...@honeybadger-io/nextjs@5.7.0) (2023-07-25)