+++ packages/react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.3](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/react@6.1.2...@honeybadger-io/react@6.1.3) (2023-08-21)
+
+**Note:** Version bump only for package @honeybadger-io/react
+
+
+
+
+
+## [6.1.2](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/react@6.1.1...@honeybadger-io/react@6.1.2) (2023-08-07)
+
+**Note:** Version bump only for package @honeybadger-io/react
+
+
+
+
+
 ## [6.1.1](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/react@6.1.0...@honeybadger-io/react@6.1.1) (2023-07-25)
 
 **Note:** Version bump only for package @honeybadger-io/react